+++ src/pages/config.tsx
@@ -1,30 +1,21 @@
-import type { FC } from 'react'
 import { Layout } from '../components/layout'
 import { defaultConfig, useConfig } from '../config/use-config'
-import { MAX_VOL_THRESHOLD, STEP_RANGE, testid } from '../constant'
+import { testid } from '../constant'
 import { dict } from '../dict'
 import { Button } from '../primitives/button'
 import { Input } from '../primitives/input'
 import { Label } from '../primitives/label'
-import { Slider } from '../primitives/slider'
 import { Switch } from '../primitives/switch'
 import { H3 } from '../primitives/typography'
+import { VolumeThreshold } from './config/volume-threshold'
 
-export const Config: FC = () => {
+export const Config = () => {
 	const [config, updateConfig] = useConfig()
 
 	const handleChange =
 		(type: keyof typeof config) => (e: React.ChangeEvent<HTMLInputElement>) => {
 			updateConfig({ [type]: e.currentTarget.value })
 		}
-
-	const handleMaxVolumeChange = (value: number[]) => {
-		updateConfig({ maxVolume: value[0] })
-	}
-
-	const handleStepVolumeChange = (value: number[]) => {
-		updateConfig({ stepVolume: value[0] })
-	}
 
 	const handleConfigDetect = () => {
 		updateConfig({
@@ -104,37 +95,7 @@ export const Config: FC = () => {
 				</section>
 
 				<section>
-					<H3>Volume threshold</H3>
-					<div className='flex flex-col gap-4'>
-						<div className='flex items-center gap-2'>
-							<Label className='w-32 shrink-0'>Max volume</Label>
-							<Slider
-								data-testid={testid.inputMaxVolume}
-								value={[config.maxVolume]}
-								min={MAX_VOL_THRESHOLD.min}
-								max={MAX_VOL_THRESHOLD.max}
-								step={MAX_VOL_THRESHOLD.step}
-								onValueChange={handleMaxVolumeChange}
-							/>
-							<span className='w-10 text-right tabular-nums text-sm'>
-								{config.maxVolume}
-							</span>
-						</div>
-						<div className='flex items-center gap-2'>
-							<Label className='w-32 shrink-0'>Step volume</Label>
-							<Slider
-								data-testid={testid.inputStepVolume}
-								value={[config.stepVolume]}
-								min={STEP_RANGE.min}
-								max={STEP_RANGE.max}
-								step={1}
-								onValueChange={handleStepVolumeChange}
-							/>
-							<span className='w-10 text-right tabular-nums text-sm'>
-								{config.stepVolume}
-							</span>
-						</div>
-					</div>
+					<VolumeThreshold />
 				</section>
 			</div>
 		</Layout>
